requirements.txt
beautifulsoup4==4.9.1
python-dotenv==0.10.1
python-telegram-bot==11.1.0
selenium==3.141.0
webdriver-manager==3.2.2
python-environ==0.4.54
